ScopeGuard as initially described by Alexandrescu and Marginean copies the callback function parameters to the guard implementation.  This means that the callback is called with the values given when make_guard gets called--not whatever the values might be when execution in the scope completes.
This changes OpenVRML's implementation of ScopeGuard to keep const references to the parameters in the guard implementation.  Also, use of boost::multi_index::detail::scope_guard (where use of OpenVRML's ScopeGuard implementation is inconvenient) now explicitly gives a reference template parameter to the instantiation of make_guard to circumvent this problem.